Move update functions into update.rs, have infer_type_ann_rec() set inferred_type
